fix: export option types
To enable reuse of the types, export them from the index file.

Also removes some unused types that were not exposed publicly.
